+++ modules/image/test/ifm3d-image-tests.cpp
@@ -542,6 +542,67 @@ TEST(Image, TimeStamp)
   EXPECT_GT(tdiff, 20);
 }
 
+TEST(Image, TimeStamps)
+{
+  std::string json =
+    R"(
+        {
+          "o3d3xx":
+          {
+            "Device":
+            {
+              "ActiveApplication": "1"
+            },
+            "Apps":
+            [
+              {
+                "TriggerMode": "1",
+                "Index": "1",
+                "Imager":
+                {
+                    "ExposureTime": "5000",
+                    "ExposureTimeList": "125;5000",
+                    "ExposureTimeRatio": "40",
+                    "Type":"under5m_moderate"
+                }
+              }
+           ]
+          }
+        }
+      )";
+
+  ifm3d::Camera::Ptr cam = std::make_shared<ifm3d::Camera>();
+  cam->FromJSON(nlohmann::json::parse(json));
+
+  ifm3d::ImageBuffer::Ptr img = std::make_shared<ifm3d::ImageBuffer>();
+  ifm3d::FrameGrabber::Ptr fg =
+    std::make_shared<ifm3d::FrameGrabber>(cam,
+                                          ifm3d::IMG_AMP | ifm3d::IMG_CART);
+
+  std::array<std::vector<ifm3d::TimePointT>, 2> tps;
+  // get two consecutive timestamps
+  for (auto& t : tps)
+    {
+      EXPECT_TRUE(fg->WaitForFrame(img.get(), 1000));
+      t = img->TimeStamps();
+    }
+
+  // checking for position 0 (last phase capture timestamp)
+  EXPECT_LT(tps[0][0], tps[1][0]);
+  auto tdiff = std::chrono::duration_cast<std::chrono::milliseconds>(
+                 tps[1][0] - tps[0][0])
+                 .count();
+  EXPECT_GT(tdiff, 20);
+
+  // the first time point need to be smaller than the second one
+  // checking for position 1 (timestamp while sending data on ethernet)
+  EXPECT_LT(tps[0][1], tps[1][1]);
+  tdiff = std::chrono::duration_cast<std::chrono::milliseconds>(tps[1][1] -
+                                                                tps[0][1])
+            .count();
+  EXPECT_GT(tdiff, 20);
+}
+
 TEST(Image, IlluTemp)
 {
   ifm3d::Camera::Ptr cam = std::make_shared<ifm3d::Camera>();
@@ -587,3 +648,5 @@ TEST(Image, DistanceNoiseImage)
 
   EXPECT_TRUE(dist_noise.type() == CV_16UC1);
 }
+
+